+++ Solutions/Digital Shadows/Data Connectors/Digital Shadows/DigitalShadowsConnectorAzureFunction/function.json
@@ -1,0 +1,12 @@
+{
+  "scriptFile": "__init__.py",
+  "bindings": [
+    {
+      "name": "mytimer",
+      "type": "timerTrigger",
+      "runOnStartup": false,
+      "direction": "in",
+      "schedule": "0 */10 * * * *"
+    }
+  ]
+}
