+++ Solutions/Digital Shadows/Data Connectors/Digital Shadows/proxies.json
@@ -1,0 +1,4 @@
+{
+  "$schema": "http://json.schemastore.org/proxies",
+  "proxies": {}
+}
