+++ Solutions/Digital Shadows/Data Connectors/Digital Shadows/requirements.txt
@@ -1,0 +1,8 @@
+# DO NOT include azure-functions-worker in this file
+# The Python Worker is managed by Azure Functions platform
+# Manually managing azure-functions-worker may cause unexpected issues
+
+azure-functions
+azure-storage-file-share
+requests
+urllib
