--- Solutions/Digital Shadows/Data Connectors/Digital shadows/DigitalShadowsConnectorAzureFunction/function.json
@@ -1,12 +1,0 @@
-{
-  "scriptFile": "__init__.py",
-  "bindings": [
-    {
-      "name": "mytimer",
-      "type": "timerTrigger",
-      "runOnStartup": false,
-      "direction": "in",
-      "schedule": "0 */10 * * * *"
-    }
-  ]
-}
